Reorganizes optimization tests.
